# functions/docs/swagger.yaml
openapi: "3.0.0"
info:
  title: "전기안전 - Firebase Functions API"
  version: "1.0.0"
  description: |
    ## 🚀 Endpoint
    배포가 완료되면 아래 주소에서 API를 확인할 수 있습니다.

    - `https://us-central1-electrical-safety-4c9bd.cloudfunctions.net/api`
      <br>  <br>
    - `https://api-2hj4xnf47q-uc.a.run.app`
paths:
...
  /device:
    get:
      summary: 센서 목록 조회
      description: 전체 센서 목록을 조회합니다.
      tags:
        - Sensors
      responses:
        "200":
          description: 센서 목록 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "센서 목록 조회 성공"
                  devices:
                    type: array
                    items:
                      type: object
                      properties:
                        sensorId:
                          type: string
                          example: "ABCD"
                          description: 센서 ID
                        address:
                          type: string
                          example: "서울시 강남구 테헤란로 123"
                          description: 센서 설치 주소
                        xCoord:
                          type: number
                          example: 37.4979
                          description: X좌표 (위도)
                        yCoord:
                          type: number
                          example: 127.0276
                          description: Y좌표 (경도)
                        createdAt:
                          type: string
                          example: "1741337167407"
                          description: 센서 등록 일시
        "500":
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                type: string
                example: "서버 내부 오류"

    post:
      summary: 새 센서 등록
      description: 센서 정보를 등록합니다.
      tags:
        - Sensors
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sensorId
                - address
                - xCoord
                - yCoord
              properties:
                sensorId:
                  type: string
                  example: "ABCD"
                  description: 센서 ID
                address:
                  type: string
                  example: "서울시 강남구 테헤란로 123"
                  description: 센서 설치 주소
                xCoord:
                  type: number
                  example: 37.4979
                  description: X좌표 (위도)
                yCoord:
                  type: number
                  example: 127.0276
                  description: Y좌표 (경도)
      responses:
        "201":
          description: 센서 등록 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "센서가 성공적으로 등록되었습니다."
                  sensorId:
                    type: string
                    example: "sensor001"
        "400":
          description: 잘못된 요청 (필수 입력값 누락)
          content:
            application/json:
              schema:
                type: string
                example: "필수 입력값이 누락되었습니다."
        "500":
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                type: string
                example: "서버 내부 오류"
  /device/{sensorId}:
    delete:
      summary: 특정 센서 삭제
      description: 센서 ID를 이용하여 해당 센서를 삭제합니다.
      tags:
        - Sensors
      parameters:
        - name: sensorId
          in: path
          required: true
          schema:
            type: string
          description: 삭제할 센서 ID
      responses:
        "200":
          description: 센서 삭제 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "센서가 성공적으로 삭제됨."
        "500":
          description: 서버 내부 오류
          content:
            application/json:
              schema:
                type: string
                example: "서버 내부 오류"
...